Allow very large S3 Objects

+++ Sources/tinys3/model/S3Object.swift
@@ -2,7 +2,7 @@ import Foundation
 
 public struct S3Object: Equatable {
     public let key: String
-    public let size: Int
+    public let size: Int64
     public let eTag: String
     public let lastModifiedAt: Date
     public let storageClass: String
